+++ NEWS.md
@@ -1,5 +1,18 @@
 # Changes to the `trajr` package
 
+## trajr 1.5.0
+
+* Added some basic 3D trajectory functionality; see `Traj3DFromCoords` and all functions prefixed with `Traj3D`.
+* `TrajAngles` now returns `NA` for the angles before and after zero-length segments. Thanks to Valentin Baumann for identifying this bug. This change could potential cause problems with existing code, but such a problem would indicate a problem with the old analysis. The bug occurred because in base R, `Arg(0) == 0`, which is not correct. Various function that take means of TrajAngles now remove NA values.
+* Added function `TrajFromTrjPoints` to simplify operations such as removing zero-length segments from a trajectory.
+* Fix documentation of `TrajDirectionalChange` and `TrajAngles`. The documentation incorrectly stated that the number of angles returned was number of points - 1. It now states that the number of angles is number of points - 2 (which is number of segments - 1).
+* Fix bug in `TrajFromCoords` that incorrectly treated NA values in columns other than coordinates or time as an error.
+* Fix bug in `TrajScale`; `yScale` was being ignored.
+* Add some more named parameters to `plot.Trajectory` to avoid passing them to `graphics::lines`. 
+* Add arg `simConstantSpeed` to `TrajRediscretize` and `Traj3DRediscretize`.
+* Add arg `check.names` to `TrajsMergeStats`.
+* Updated github continuous integration actions (using usethis::use_github_action("test-coverage") and usethis::use_github_action_check_release()). This change does not affect users
+
 ## trajr 1.4.0
 
 * Allow the creation of trajectories with 0 points in `TrajFromCoords`.